skins2: remove questionable trick
This trick only works if the action ends up in the state really being toggled.
This is not always the case and can then lead to undesirable effects,
for instance the play button going to pause when no playback is underway.

This trick was needed to prevent visual artefacts when refreshing was a heavy
task, but is no longer needed since refreshing has been heavily optimised.