macosx: fixed media key support on OS X 10.6+
Note that this rework changes the current behavior: if VLC is in the background, the keys still control it, however if you start another key user (such as iTunes, QTP, etc.), it won't interfere and control the other app. That's why macosx-mediakeys-background was removed.

This commits also adds a bunch of missing delegate declarations

Thanks to Spotify for releasing their code.